fix: jsx 방식 조건부 렌더링으로 변경
