+++ src/components/Error/Error.jsx
@@ -1,0 +1,3 @@
+export default function Error({ errorMessage }) {
+  return <div>{errorMessage}</div>;
+}
